requirements.txt
affine==2.4.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.13
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
awscli==1.16.201
Babel==2.12.1
backcall==0.2.0
billiard==4.1.0
botocore==1.12.191
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colorama==0.3.9
comm==0.1.3
configobj==5.0.8
contextily==1.3.0
contourpy==1.1.0
coverage==7.2.7
cryptography==41.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.1
distro==1.8.0
docutils==0.20.1
dpath==2.1.6
dulwich==0.21.5
dvc==3.10.1
dvc-data==2.8.1
dvc-http==2.30.2
dvc-objects==0.24.1
dvc-render==0.5.3
dvc-studio-client==0.11.0
dvc-task==0.3.0
dvclive==2.13.0
executing==1.2.0
filelock==3.12.2
Fiona==1.9.4.post1
flake8==6.1.0
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
funcy==2.0
geographiclib==2.0
geopandas==0.13.2
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.32
grandalf==0.8
hydra-core==1.3.2
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.25.0
ipython==8.14.0
iterative-telemetry==0.0.8
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
kombu==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mercantile==1.2.1
multidict==6.0.4
nanotime==0.5.2
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.1
omegaconf==2.3.0
orjson==3.9.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydot==1.4.2
pyflakes==3.1.0
pygit2==1.12.2
Pygments==2.15.1
pygtrie==2.5.0
pyparsing==3.0.9
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.1
pyzmq==25.1.0
rasterio==1.3.8
requests==2.31.0
rich==13.5.0
rsa==3.4.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.2.1
scikit-learn==1.3.0
scipy==1.11.1
scmrepo==1.1.0
seaborn==0.12.2
shapely==2.0.1
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
snowballstemmer==2.2.0
snuggs==1.4.7
Sphinx==7.1.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqltrie==0.7.0
# Editable Git install with no remote (src==0.1.0)
-e /Users/shubham/give_directly
stack-data==0.6.2
tabulate==0.9.0
threadpoolctl==3.2.0
tomlkit==0.12.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.25.11
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
xyzservices==2023.7.0
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.16.2
